25-05-2018

+++ database/migrations/2018_05_25_041330_create_orders_table.php
@@ -17,7 +17,7 @@ class CreateOrdersTable extends Migration
             $table->increments('id');
             $table->string('shop_url');
             $table->string('account_key');
-            $table->integer('order_id');
+            $table->bigInteger('order_id');
         });
     }
